Register is now a modal pop-up
Close #97
