+++ package.json
@@ -26,7 +26,8 @@
     "solid-auth-cli": "^1.0.10",
     "solid-auth-client": "^2.4.1",
     "solid-file-client": "^1.0.2",
-    "styled-components": "^4.1.3"
+    "styled-components": "^4.1.3",
+    "wrld.js": "^0.1.2179"
   },
   "scripts": {
     "start": "react-scripts start",
