+++ src/components/NavBar/Register/ProviderItem/index.js
