+++ src/components/NavBar/Register/index.js
@@ -1,0 +1,3 @@
+import RegisterModal from "./register-modal.component";
+
+export default RegisterModal;
