+++ src/components/NavBar/Register/register.container.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import RegisterComponent from './register.component';
-import {Provider} from '../../services';
+import {Provider} from '../../../services';
 
 type Props = {};
