--- src/containers/Register/index.js
@@ -1,3 +1,0 @@
-import RegisterContainer from "./register.container";
-
-export default RegisterContainer ;
