+++ src/routes.js
@@ -4,17 +4,15 @@ import {BrowserRouter as Router, Switch} from "react-router-dom";
 import {
 	Friends,
 	Home,
-	Login,
 	MyRoutes,
 	NotFound,
 	Profile,
-	Register,
 	RoutesList,
 	ShowRoute,
 	Upload
 } from "./containers";
 
-import {NotLoggedInLayout, PrivateLayout, PublicLayout} from "./layouts";
+import {PrivateLayout, PublicLayout} from "./layouts";
 
 
 /**
@@ -23,29 +21,27 @@ import {NotLoggedInLayout, PrivateLayout, PublicLayout} from "./layouts";
  * @constructor
  */
 const Routes = () => {
-  return (
-  <Router>
-    <Fragment>
-      {/* Chooses the first route matching the direction and loads it */}
-      <Switch>
-        {
-          /*
-            PublicLayout para rutas que se tienen que mostrar a todos los usuarios(estén o no loggeados)
-            NotLoggedInLayout para rutas que se tienen que mostrar a usuarios que NO están loggeados
-            PrivateLayout para rutas que se tienen que mostrar a usuarios que SI están loggeados
-          */
-        }
-        <PublicLayout component={Home} path="/" exact />      {/* Homepage - "/" */}
-        <NotLoggedInLayout component={Login} path="/login" exact /> {/* Login - "/login" */}
-        <NotLoggedInLayout component={Register} path="/register" exact /> {/* Register - "/register" */}
-        <PrivateLayout component={Friends} path="/friends" exact /> {/* Friends - "/friends" */}
-		<PrivateLayout component={ShowRoute} path="/route" exact/> {/* ShowRoute - "/route" */}
-        <PrivateLayout component={Profile} path="/profile" exact /> {/* Profile - "/profile" */}
-        <PrivateLayout component={MyRoutes} path="/myRoutes" exact /> {/* My Routes - "/myROutes" */}
-        <PrivateLayout component={RoutesList} path="/routesList" exact /> {/* Routes List - "/routesList" */}
-        <PrivateLayout component={ShowRoute} path="/showRoute" exact /> {/* View Map - "/viewMap" */}
-        <PrivateLayout component={Upload} path="/upload" exact /> {/* Upload- "/upload" */}
-        <PublicLayout component={NotFound} path="*"/> {/* Error - "*" */}
+	return (
+		<Router>
+			<Fragment>
+				{/* Chooses the first route matching the direction and loads it */}
+				<Switch>
+					{
+						/*
+						  PublicLayout para rutas que se tienen que mostrar a todos los usuarios(estén o no loggeados)
+						  NotLoggedInLayout para rutas que se tienen que mostrar a usuarios que NO están loggeados
+						  PrivateLayout para rutas que se tienen que mostrar a usuarios que SI están loggeados
+						*/
+					}
+					<PublicLayout component={Home} path="/" exact/> {/* Homepage - "/" */}
+					<PrivateLayout component={Friends} path="/friends" exact/> {/* Friends - "/friends" */}
+					<PrivateLayout component={ShowRoute} path="/route" exact/> {/* ShowRoute - "/route" */}
+					<PrivateLayout component={Profile} path="/profile" exact/> {/* Profile - "/profile" */}
+					<PrivateLayout component={MyRoutes} path="/myRoutes" exact/> {/* My Routes - "/myROutes" */}
+					<PrivateLayout component={RoutesList} path="/routesList" exact/> {/* Routes List - "/routesList" */}
+					<PrivateLayout component={ShowRoute} path="/showRoute" exact/> {/* View Map - "/viewMap" */}
+					<PrivateLayout component={Upload} path="/upload" exact/> {/* Upload- "/upload" */}
+					<PublicLayout component={NotFound} path="*"/> {/* Error - "*" */}
 
 				</Switch>
 			</Fragment>
